+++ vite-vue3/package.json
@@ -16,7 +16,9 @@
     "bootstrap": "^5.1.3",
     "sass": "^1.43.2",
     "sweetalert2": "^11.1.9",
-    "vue": "^3.2.16"
+    "vue": "^3.2.16",
+    "vue-router": "^4.0.12",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^1.9.3",
