--- vite-vue3/src/_main.scss
@@ -1,3 +1,0 @@
-@import "./assets/scss/common";
-@import "bootstrap/scss/bootstrap";
-@import "sweetalert2/src/sweetalert2.scss";
