+++ vite-vue3/src/components/MainFooter.vue
@@ -1,0 +1,17 @@
+<script setup lang="ts">
+import { ref, reactive } from 'vue'
+
+defineProps<{ msg: string }>()
+
+const count = ref(0)
+</script>
+
+<template>
+  <footer class="footer">
+      
+  </footer>
+</template>
+
+<style lang="scss" scoped>
+
+</style>
