+++ vite-vue3/src/main.ts
@@ -1,8 +1,10 @@
 import { createApp } from 'vue';
 import App from './App.vue';
 
-// 공통 SCSS(라이브러리 SCSS 포함)
-import './_main.scss';
+/* https://itnext.io/vue-tricks-smart-layouts-for-vuejs-5c61a472b69b
+ * 레이아웃 만드는 법 설명 잘나와 있음
+ * 레이아웃 잘 활용하면 생산성 높아짐
+ */
 
 // 부트스트랩5
 import 'bootstrap';
@@ -19,6 +21,7 @@ import { FontAwesomeIcon, FontAwesomeLayers, FontAwesomeLayersText } from '@fort
 library.add(fas, fab, far);
 
 createApp(App)
+    // 폰트 어썸
     .component('font-awesome-icon', FontAwesomeIcon)
     .component('font-awesome-layers', FontAwesomeLayers)
     .component('font-awesome-layers-text', FontAwesomeLayersText)
